Weather Application

+++ app/src/main/java/com/example/weatherapplication/Model/Main.java
@@ -5,7 +5,7 @@ public class Main {
     private double feels_like;
     private double temp_min;
     private double temp_max;
-    private int pressure;
+    private float pressure;
     private int humidity;
 
     public Main() {
@@ -43,11 +43,11 @@ public class Main {
         this.temp_max = temp_max;
     }
 
-    public int getPressure() {
+    public float getPressure() {
         return pressure;
     }
 
-    public void setPressure(int pressure) {
+    public void setPressure(float pressure) {
         this.pressure = pressure;
     }
